Skip monthly counters deletion with option --skip-monthly

+++ onadata/apps/logger/management/commands/populate_submission_counters.py
@@ -51,7 +51,7 @@ class Command(BaseCommand):
         )
 
         parser.add_argument(
-            '--skip_monthly',
+            '--skip-monthly',
             action='store_true',
             default=False,
             help='Skip updating monthly counters. Default is False',
@@ -185,14 +185,15 @@ class Command(BaseCommand):
         # Because we don't have a real date field on `MonthlyXFormSubmissionCounter`
         # but we need to cast `year` and `month` as a date field to
         # compare it with `self._date_threshold`
-        MonthlyXFormSubmissionCounter.objects.annotate(
-            date=Cast(
-                Concat(
-                    F('year'), Value('-'), F('month'), Value('-'), 1
-                ),
-                DateField(),
-            )
-        ).filter(user_id=user.pk, date__gte=self._date_threshold).delete()
+        if not self._skip_monthly:
+            MonthlyXFormSubmissionCounter.objects.annotate(
+                date=Cast(
+                    Concat(
+                        F('year'), Value('-'), F('month'), Value('-'), 1
+                    ),
+                    DateField(),
+                )
+            ).filter(user_id=user.pk, date__gte=self._date_threshold).delete()
 
     def suspend_submissions_for_user(self, user: 'auth.User'):
         # Retrieve or create user's profile.
